+++ sdk/communication/azure-communication-callautomation/assets.json
@@ -1,0 +1,6 @@
+{
+  "AssetsRepo": "Azure/azure-sdk-assets",
+  "AssetsRepoPrefixPath": "java",
+  "TagPrefix": "java/communication/azure-communication-callautomation",
+  "Tag": "java/communication/azure-communication-callautomation_6c901370a0"
+}
