+++ sdk/communication/azure-communication-callautomation/pom.xml
@@ -77,7 +77,7 @@
     <dependency>
       <groupId>com.azure</groupId>
       <artifactId>azure-core-test</artifactId>
-      <version>1.19.0</version> <!-- {x-version-update;com.azure:azure-core-test;dependency} -->
+      <version>1.20.0-beta.1</version> <!-- {x-version-update;com.azure:azure-core-test;dependency} -->
       <scope>test</scope>
     </dependency>
     <dependency>
