--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallAutomationAsyncClientAutomatedLiveTests.createVOIPCallAndAnswerThenHangupAutomatedTest[1].json
@@ -1,140 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "5f72fbb2-703e-48e3-8225-cc7786ab573d",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:54 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTQgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "3oH7FafeeEW7HQfyPgU2ag.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAACghdLQQCMvQJd2Cb8gWYenV1NURURHRTAxMjEAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "5f72fbb2-703e-48e3-8225-cc7786ab573d",
-        "X-Processing-Time": "35ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "abcd684a-f5ad-44c2-9e29-aa5d431c6cf7",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:55 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTUgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "9ckMilIqP0SEBWzPDB5sQA.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAADY9pUwmefWQrhJ93s9L08FV1NURURHRTA4MDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "abcd684a-f5ad-44c2-9e29-aa5d431c6cf7",
-        "X-Processing-Time": "16ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/calling/callConnections?api-version=2023-01-15-preview",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "572",
-        "Content-Type": "application/json",
-        "Repeatability-First-Sent": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "Repeatability-Request-ID": "f4c77452-3462-4741-a1c8-e9723010e923",
-        "User-Agent": "azsdk-java-azure-communication-callautomation/1.1.0-beta.1 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "ebb04d63-5849-4aa5-b492-1031c2d4899b",
-        "x-ms-content-sha256": "7EyW\u002BZt1q7aXLOwVa\u002BxWybEKJ0gs9YiKLg0VKQNj6w8=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovY2FsbGluZy9jYWxsQ29ubmVjdGlvbnM/YXBpLXZlcnNpb249MjAyMy0wMS0xNS1wcmV2aWV3ClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTggR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tOzdFeVcrWnQxcTdhWExPd1ZhK3hXeWJFS0owZ3M5WWlLTGcwVktRTmo2dzg9"
-      },
-      "RequestBody": {
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264"
-            }
-          }
-        ],
-        "sourceIdentity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca"
-        },
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264"
-      },
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "Content-Length": "970",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:38:01 GMT",
-        "X-Azure-Ref": "0ueqsZAAAAADwByuXWoBwR78kT6ZIwSoBV1NURURHRTAxMjEAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Microsoft-Skype-Chain-ID": "2d6f7310-531a-4aa5-abf6-a8c577bcadbb",
-        "x-ms-client-request-id": "ebb04d63-5849-4aa5-b492-1031c2d4899b"
-      },
-      "ResponseBody": {
-        "callConnectionId": "401f0b00-0762-413d-9812-13fe6f663cbf",
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264",
-            "kind": "communicationUser",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264"
-            }
-          }
-        ],
-        "callConnectionState": "connecting",
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a5d-85f4-343a0d00a264",
-        "mediaSubscriptionId": "496f9205-b502-4076-af55-7bd827ad37df",
-        "sourceDisplayName": "",
-        "sourceIdentity": {
-          "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca",
-          "kind": "communicationUser",
-          "communicationUser": {
-            "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79ff-78fe-343a0d0088ca"
-          }
-        },
-        "correlationId": "2d6f7310-531a-4aa5-abf6-a8c577bcadbb"
-      }
-    }
-  ],
-  "Variables": {}
-}
