--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallAutomationAsyncClientAutomatedLiveTests.createVOIPCallAndRejectAutomatedTest[1].json
@@ -1,140 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "7f54e2d5-2fd9-4432-96be-7c959e0f76af",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:54 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTQgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "AS2T4LlWtkiWiAyhJHI0aA.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAACERG0UgRBYQolKhuyCizeuV1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "7f54e2d5-2fd9-4432-96be-7c959e0f76af",
-        "X-Processing-Time": "35ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd0"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "37fe7323-c20f-45c4-9f42-deb8dcaa829e",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:55 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTUgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "2V\u002BQgyLFCEixcNGQvy8ecw.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAABocjYcQPFrT4X5PrddtW94V1NURURHRTA4MjAAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "37fe7323-c20f-45c4-9f42-deb8dcaa829e",
-        "X-Processing-Time": "18ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/calling/callConnections?api-version=2023-01-15-preview",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "572",
-        "Content-Type": "application/json",
-        "Repeatability-First-Sent": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "Repeatability-Request-ID": "e8d8052a-9419-40ed-be35-a80cd572ca8a",
-        "User-Agent": "azsdk-java-azure-communication-callautomation/1.1.0-beta.1 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "83439a9e-ab7e-481f-8653-5013ea48b59e",
-        "x-ms-content-sha256": "EVATDlLFSWW3omsFbasd0fwN2\u002B/2CLfNT6G1pUlXnI0=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovY2FsbGluZy9jYWxsQ29ubmVjdGlvbnM/YXBpLXZlcnNpb249MjAyMy0wMS0xNS1wcmV2aWV3ClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTggR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO0VWQVREbExGU1dXM29tc0ZiYXNkMGZ3TjIrLzJDTGZOVDZHMXBVbFhuSTA9"
-      },
-      "RequestBody": {
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3"
-            }
-          }
-        ],
-        "sourceIdentity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd0"
-        },
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd08acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3"
-      },
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "Content-Length": "970",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:38:01 GMT",
-        "X-Azure-Ref": "0ueqsZAAAAAAQu0B6xVnHSpSrnYW4PJO4V1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Microsoft-Skype-Chain-ID": "08490c67-2491-4b27-a547-d305d6d9cabb",
-        "x-ms-client-request-id": "83439a9e-ab7e-481f-8653-5013ea48b59e"
-      },
-      "ResponseBody": {
-        "callConnectionId": "411f0b00-66cb-4802-aa48-b782dec178f5",
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3",
-            "kind": "communicationUser",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3"
-            }
-          }
-        ],
-        "callConnectionState": "connecting",
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd08acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-28f4-343a0d0077c3",
-        "mediaSubscriptionId": "f671df8a-cc5d-4d77-a8a5-1c5a5b49c684",
-        "sourceDisplayName": "",
-        "sourceIdentity": {
-          "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd0",
-          "kind": "communicationUser",
-          "communicationUser": {
-            "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-4ff7-343a0d007fd0"
-          }
-        },
-        "correlationId": "08490c67-2491-4b27-a547-d305d6d9cabb"
-      }
-    }
-  ],
-  "Variables": {}
-}
