--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallMediaAsyncAutomatedLiveTests.dtmfActionsInACallAutomatedTest[1].json
@@ -1,4 +1,0 @@
-{
-  "Entries": [],
-  "Variables": {}
-}
