--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallMediaAsyncAutomatedLiveTests.playMediaInACallAutomatedTest[1].json
@@ -1,140 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "3994b835-0e05-491a-acc1-5e4ed2b511bf",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:54 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTQgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "pTiToV06JUqdAE7b/dPwZw.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAABrDiJsXZWTQ6\u002B7ZwYzFRKoV1NURURHRTA4MDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "3994b835-0e05-491a-acc1-5e4ed2b511bf",
-        "X-Processing-Time": "33ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a262"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "9dabf173-81b9-4042-8a6e-7ce85a1f865f",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:55 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTUgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "uUdk9Xn9FU22arblLzejsQ.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAACa46vpjPmzTqp0Gbxuv7ymV1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "9dabf173-81b9-4042-8a6e-7ce85a1f865f",
-        "X-Processing-Time": "16ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/calling/callConnections?api-version=2023-01-15-preview",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "572",
-        "Content-Type": "application/json",
-        "Repeatability-First-Sent": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "Repeatability-Request-ID": "4cd5eff5-d664-4d9a-8592-bc2dc8baa3dd",
-        "User-Agent": "azsdk-java-azure-communication-callautomation/1.1.0-beta.1 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "0c12cf5a-8950-4b04-9558-f975db46dc40",
-        "x-ms-content-sha256": "VJtXBWptlbcpj8kwnaJyw4mhDRdX4c7NT\u002BYDLJvZDJw=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovY2FsbGluZy9jYWxsQ29ubmVjdGlvbnM/YXBpLXZlcnNpb249MjAyMy0wMS0xNS1wcmV2aWV3ClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTggR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1ZKdFhCV3B0bGJjcGo4a3duYUp5dzRtaERSZFg0YzdOVCtZRExKdlpESnc9"
-      },
-      "RequestBody": {
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1"
-            }
-          }
-        ],
-        "sourceIdentity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a262"
-        },
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a2628acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1"
-      },
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "Content-Length": "970",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:38:01 GMT",
-        "X-Azure-Ref": "0ueqsZAAAAACDHC4R4dq4S4ZmhjsHHBlcV1NURURHRTA4MjAAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Microsoft-Skype-Chain-ID": "05c344ed-24b1-4f6f-9952-9de50d74ca2c",
-        "x-ms-client-request-id": "0c12cf5a-8950-4b04-9558-f975db46dc40"
-      },
-      "ResponseBody": {
-        "callConnectionId": "401f0b00-6d9f-4c99-84dd-6837a043ca1f",
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1",
-            "kind": "communicationUser",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1"
-            }
-          }
-        ],
-        "callConnectionState": "connecting",
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a2628acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a63-4ff7-343a0d007fd1",
-        "mediaSubscriptionId": "28e6e4f6-1764-4a92-8c8c-451654db7b75",
-        "sourceDisplayName": "",
-        "sourceIdentity": {
-          "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a262",
-          "kind": "communicationUser",
-          "communicationUser": {
-            "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79f4-85f4-343a0d00a262"
-          }
-        },
-        "correlationId": "05c344ed-24b1-4f6f-9952-9de50d74ca2c"
-      }
-    }
-  ],
-  "Variables": {}
-}
