--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallRecordingAutomatedLiveTests.createACSCallAndUnmixedAudioTest[1].json
@@ -1,140 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "f757a50f-ba5b-4d76-8f94-6bfc3b2bca02",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:54 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTQgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "mfIpHpF/5kOPtaACzJmvvg.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAADsd0N9m/3vQYY1kCGkitz1V1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "f757a50f-ba5b-4d76-8f94-6bfc3b2bca02",
-        "X-Processing-Time": "17ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a263"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "f4aa7be2-0eb5-4ce4-ae70-7b1b980d5960",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:55 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTUgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "m1W5b0hJXESUgNRKYjl1cw.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAAD5/ZwLPPYTS4tMSNTt9/0OV1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "f4aa7be2-0eb5-4ce4-ae70-7b1b980d5960",
-        "X-Processing-Time": "16ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/calling/callConnections?api-version=2023-01-15-preview",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "572",
-        "Content-Type": "application/json",
-        "Repeatability-First-Sent": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "Repeatability-Request-ID": "58c4789e-8a4e-4791-9965-18760f7d930d",
-        "User-Agent": "azsdk-java-azure-communication-callautomation/1.1.0-beta.1 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "d6bd59b5-e6dd-4345-8203-a5ad78b98e60",
-        "x-ms-content-sha256": "NYMdXo8MPiGuai6QYti5zvM3rc44mqAi838uYtVHktE=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovY2FsbGluZy9jYWxsQ29ubmVjdGlvbnM/YXBpLXZlcnNpb249MjAyMy0wMS0xNS1wcmV2aWV3ClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTggR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO05ZTWRYbzhNUGlHdWFpNlFZdGk1enZNM3JjNDRtcUFpODM4dVl0VkhrdEU9"
-      },
-      "RequestBody": {
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265"
-            }
-          }
-        ],
-        "sourceIdentity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a263"
-        },
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a2638acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265"
-      },
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "Content-Length": "970",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:38:01 GMT",
-        "X-Azure-Ref": "0ueqsZAAAAACuNtuP71UTR7mFCIbauEFOV1NURURHRTAxMDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Microsoft-Skype-Chain-ID": "0d749c51-21bd-4c68-af02-e2151f47cb73",
-        "x-ms-client-request-id": "d6bd59b5-e6dd-4345-8203-a5ad78b98e60"
-      },
-      "ResponseBody": {
-        "callConnectionId": "401f0b00-0689-4958-9f49-5c71ab35ea51",
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265",
-            "kind": "communicationUser",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265"
-            }
-          }
-        ],
-        "callConnectionState": "connecting",
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a2638acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a85-85f4-343a0d00a265",
-        "mediaSubscriptionId": "3e486d81-8ff3-45e7-ba73-4aab5708d76d",
-        "sourceDisplayName": "",
-        "sourceIdentity": {
-          "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a263",
-          "kind": "communicationUser",
-          "communicationUser": {
-            "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a43-85f4-343a0d00a263"
-          }
-        },
-        "correlationId": "0d749c51-21bd-4c68-af02-e2151f47cb73"
-      }
-    }
-  ],
-  "Variables": {}
-}
