--- sdk/communication/azure-communication-callautomation/src/test/resources/session-records/CallRecordingAutomatedLiveTests.createACSCallUnmixedAudioAffinityTest[1].json
@@ -1,140 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "5833674a-972e-4667-8a7a-401522af97d8",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:54 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTQgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "XLpaZRoZF0K5ClaQ9bE2nw.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAADgaBRjaelVQqKZipIlKqmcV1NURURHRTA4MjAAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "5833674a-972e-4667-8a7a-401522af97d8",
-        "X-Processing-Time": "29ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c2"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/identities?api-version=2022-10-01",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "2",
-        "Content-Type": "application/json",
-        "User-Agent": "azsdk-java-azure-communication-identity/1.4.7 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "610d519c-31e0-4ffe-a45f-6c169005ec53",
-        "x-ms-content-sha256": "RBNvo1WzZ4oRRq0W9\u002BhknpT7T8If536DEMBg9hyq/4o=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:55 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovaWRlbnRpdGllcz9hcGktdmVyc2lvbj0yMDIyLTEwLTAxClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTUgR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO1JCTnZvMVd6WjRvUlJxMFc5K2hrbnBUN1Q4SWY1MzZERU1CZzloeXEvNG89"
-      },
-      "RequestBody": {},
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "api-supported-versions": "2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-10-31-preview, 2021-11-01, 2022-06-01, 2022-10-01, 2023-08-01",
-        "Content-Length": "101",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:37:57 GMT",
-        "MS-CV": "t/a2q0SP4U6uOa4hwPjpXA.0",
-        "Request-Context": "appId=",
-        "Strict-Transport-Security": "max-age=2592000",
-        "X-Azure-Ref": "0teqsZAAAAAAM/wcq0656SaU7zkSXqp9zV1NURURHRTA4MTIAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "x-ms-client-request-id": "610d519c-31e0-4ffe-a45f-6c169005ec53",
-        "X-Processing-Time": "32ms"
-      },
-      "ResponseBody": {
-        "identity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45"
-        }
-      }
-    },
-    {
-      "RequestUri": "https://REDACTED/calling/callConnections?api-version=2023-01-15-preview",
-      "RequestMethod": "POST",
-      "RequestHeaders": {
-        "Accept": "application/json",
-        "Authorization": "Sanitized",
-        "Content-Length": "572",
-        "Content-Type": "application/json",
-        "Repeatability-First-Sent": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "Repeatability-Request-ID": "07d96b5c-d3a2-45a7-a3d0-90257e719b85",
-        "User-Agent": "azsdk-java-azure-communication-callautomation/1.1.0-beta.1 (17.0.1; Windows 11; 10.0)",
-        "x-ms-client-request-id": "f5f355ad-7823-4126-b43d-ff9832ab90b4",
-        "x-ms-content-sha256": "l3hZBKQwGUl4ta3if\u002BgHS47VuJNZbhEG6Uba9SPnueA=",
-        "x-ms-date": "Tue, 11 Jul 2023 05:37:58 GMT",
-        "x-ms-hmac-string-to-sign-base64": "UE9TVAovY2FsbGluZy9jYWxsQ29ubmVjdGlvbnM/YXBpLXZlcnNpb249MjAyMy0wMS0xNS1wcmV2aWV3ClR1ZSwgMTEgSnVsIDIwMjMgMDU6Mzc6NTggR01UO3NybmFnYXItYWNzLmNvbW11bmljYXRpb24uYXp1cmUuY29tO2wzaFpCS1F3R1VsNHRhM2lmK2dIUzQ3VnVKTlpiaEVHNlViYTlTUG51ZUE9"
-      },
-      "RequestBody": {
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45"
-            }
-          }
-        ],
-        "sourceIdentity": {
-          "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c2"
-        },
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c28acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45"
-      },
-      "StatusCode": 201,
-      "ResponseHeaders": {
-        "Content-Length": "970",
-        "Content-Type": "application/json; charset=utf-8",
-        "Date": "Tue, 11 Jul 2023 05:38:00 GMT",
-        "X-Azure-Ref": "0ueqsZAAAAAAJ0I/RgjMlQKBBdxf2dP4jV1NURURHRTA4MTIAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-        "X-Cache": "CONFIG_NOCACHE",
-        "X-Microsoft-Skype-Chain-ID": "4f4609bc-0289-480b-b5cf-28f0e4ca32b5",
-        "x-ms-client-request-id": "f5f355ad-7823-4126-b43d-ff9832ab90b4"
-      },
-      "ResponseBody": {
-        "callConnectionId": "401f0b00-0837-4a82-868c-0998a0e8994b",
-        "targets": [
-          {
-            "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45",
-            "kind": "communicationUser",
-            "communicationUser": {
-              "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45"
-            }
-          }
-        ],
-        "callConnectionState": "connecting",
-        "callbackUri": "https://incomingcalldispatcher.azurewebsites.net/api/servicebuscallback/events?q=8acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c28acs01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-7a6c-35f3-343a0d006b45",
-        "mediaSubscriptionId": "0552ac99-22c6-4c81-bb72-ed847d606640",
-        "sourceDisplayName": "",
-        "sourceIdentity": {
-          "rawId": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c2",
-          "kind": "communicationUser",
-          "communicationUser": {
-            "id": "8:acs:01bd4b25-ad44-4bc2-88ad-e67ce9593891_00000019-e356-79fb-28f4-343a0d0077c2"
-          }
-        },
-        "correlationId": "4f4609bc-0289-480b-b5cf-28f0e4ca32b5"
-      }
-    }
-  ],
-  "Variables": {}
-}
